+++ components/php_ext_wifi/php_wifi.h
@@ -1,0 +1,9 @@
+#ifndef PHP_WIFI_H
+#define PHP_WIFI_H
+
+#include "php.h"
+
+extern zend_module_entry wifi_module_entry;
+PHP_MINIT_FUNCTION(wifi);
+
+#endif /* PHP_WIFI_H */